Isolating Carousel from Uploader and Making carousel contents separate from carousel
