+++ packages/shared-components/Carousel/style.js
@@ -21,8 +21,8 @@ export const Arrow = styled.button`
 
 export const CarouselContainer = styled.div`
   display: flex;
-  padding-left: ${props => (props.editMode ? '16px' : 0)};
-  width: calc(${props => props.cardWidth + (cardMargin * 2)}px *  ${$props => $props.totalCards});
+  padding-left: ${props => (props.largeCards ? '16px' : 0)};
+  width: calc(${props => props.cardWidth + (cardMargin * 2)}px *  ${$props => $props.totalCardsToShow});
   transform: ${props => `translateX(calc(-${props.cardWidth + (cardMargin * 2)}px * ${props.selectedItem}`}));
   transition: all 0.3s ease-out;
 `;
@@ -31,9 +31,9 @@ export const CarouselCard = styled.div`
   display: flex;
   align-items: center;
   justify-content: center;
-  margin: ${props => (props.editMode ? `${cardMargin}px` : `16px ${cardMargin}px`)};
-  height: ${$props => $props.cardHeight}px;
-  width: ${$props => $props.cardWidth}px;
+  margin: ${props => (props.largeCards ? `${cardMargin}px` : `16px ${cardMargin}px`)};
+  height: ${props => props.cardHeight ? props.cardHeight : 198}px;
+  width: ${props => props.cardWidth ? props.cardWidth : 110}px;
   background-color: ${grayLighter};
   background-size: cover;
   background-position: center;
@@ -43,10 +43,5 @@ export const CarouselCard = styled.div`
 export const SliderCarousel = styled.div`
   position: relative;
   overflow: hidden;
-  margin: ${props => (props.editMode ? '0 -16px 0 -16px' : '0 16px 0 12px')};
-`;
-
-export const IconWrapper = styled.div`
-  display: flex;
-  opacity: 0.8;
+  margin: ${props => (props.largeCards ? '0 -16px 0 -16px' : '0 16px 0 12px')};
 `;
